+++ core-ng/src/main/java/core/framework/internal/web/sys/KafkaController.java
@@ -1,6 +1,6 @@
 package core.framework.internal.web.sys;
 
-import core.framework.internal.kafka.MessageHeaders;
+import core.framework.internal.kafka.KafkaMessage;
 import core.framework.internal.kafka.MessageListener;
 import core.framework.internal.kafka.MessageProcess;
 import core.framework.internal.kafka.MessageProducer;
@@ -19,6 +19,7 @@ import org.apache.kafka.common.header.Headers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
 import java.util.List;
 
 import static core.framework.log.Markers.errorCode;
@@ -59,22 +60,28 @@ public class KafkaController {
 
         @SuppressWarnings("unchecked")
         MessageProcess<Object> process = (MessageProcess<Object>) listener.processes.get(topic);
-        if (process == null) throw new Error("handler not found, topic=" + topic);
+        @SuppressWarnings("unchecked")
+        MessageProcess<Object> bulkProcess = (MessageProcess<Object>) listener.bulkProcesses.get(topic);
+        if (process != null) {
+            Object message = message(topic, key, body, process, actionLog);
+            process.handler().handle(key, message);
+        } else if (bulkProcess != null) {
+            Object message = message(topic, key, body, bulkProcess, actionLog);
+            bulkProcess.bulkHandler().handle(List.of(new Message<>(key, message)));
+        } else {
+            throw new Error("handler not found, topic=" + topic);
+        }
+        return Response.text(Strings.format("message handled, topic={}, key={}, message={}", topic, key, new String(body, UTF_8)));
+    }
+
+    private Object message(String topic, String key, byte[] body, MessageProcess<Object> process, ActionLog actionLog) throws IOException {
         Object message = process.reader.fromJSON(body);
         process.validator.validate(message, false);
         logger.debug("[message] topic={}, key={}, value={}", topic, key, new BytesLogParam(Strings.bytes(JSON.toJSON(message))));    // log converted message
-        if (process.handler != null) {
-            String handler = process.handler.getClass().getCanonicalName();
-            actionLog.context.put("handler", List.of(handler));
-            logger.debug("handler={}", handler);
-            process.handler.handle(key, message);
-        } else {
-            String handler = process.bulkHandler.getClass().getCanonicalName();
-            actionLog.context.put("handler", List.of(handler));
-            logger.debug("handler={}", handler);
-            process.bulkHandler.handle(List.of(new Message<>(key, message)));
-        }
-        return Response.text(Strings.format("message handled, topic={}, key={}, message={}", topic, key, new String(body, UTF_8)));
+        String handler = process.handler.getClass().getCanonicalName();
+        actionLog.context.put("handler", List.of(handler));
+        logger.debug("handler={}", handler);
+        return message;
     }
 
     private ActionLog initAction(String topic, String key) {
@@ -87,10 +94,10 @@ public class KafkaController {
     ProducerRecord<byte[], byte[]> record(String topic, String key, byte[] body, ActionLog actionLog) {
         var record = new ProducerRecord<>(topic, Strings.bytes(key), body);
         Headers headers = record.headers();
-        headers.add(MessageHeaders.HEADER_CLIENT, Strings.bytes(KafkaController.class.getSimpleName()));
-        headers.add(MessageHeaders.HEADER_TRACE, Strings.bytes(Trace.CASCADE.name()));  // auto trace
-        headers.add(MessageHeaders.HEADER_CORRELATION_ID, Strings.bytes(actionLog.correlationId()));
-        headers.add(MessageHeaders.HEADER_REF_ID, Strings.bytes(actionLog.id));
+        headers.add(KafkaMessage.HEADER_CLIENT, Strings.bytes(KafkaController.class.getSimpleName()));
+        headers.add(KafkaMessage.HEADER_TRACE, Strings.bytes(Trace.CASCADE.name()));  // auto trace
+        headers.add(KafkaMessage.HEADER_CORRELATION_ID, Strings.bytes(actionLog.correlationId()));
+        headers.add(KafkaMessage.HEADER_REF_ID, Strings.bytes(actionLog.id));
         return record;
     }
 }