+++ core-ng/src/test/java/core/framework/internal/kafka/MessageListenerThreadTest.java
@@ -11,7 +11,6 @@ import core.framework.kafka.MessageHandler;
 import core.framework.util.Strings;
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.common.header.internals.RecordHeaders;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -36,34 +35,26 @@ class MessageListenerThreadTest {
     @Mock
     MessageHandler<TestMessage> messageHandler;
     @Mock
-    Consumer<String, byte[]> consumer;
-    @Mock
     BulkMessageHandler<TestMessage> bulkMessageHandler;
+    @Mock
+    Consumer<String, byte[]> consumer;
     private MessageListenerThread thread;
     private LogManager logManager;
 
     @BeforeEach
-    void createKafkaMessageListenerThread() {
+    void createMessageListenerThread() {
         logManager = new LogManager();
-        thread = new MessageListenerThread("listener-thread-1", consumer, new MessageListener(null, null, logManager, 300_000L));
-    }
-
-    @Test
-    void header() {
-        var headers = new RecordHeaders();
-        headers.add("header", Strings.bytes("value"));
-        assertThat(thread.header(headers, "header")).isEqualTo("value");
-        assertThat(thread.header(headers, "nonExisted")).isNull();
+        thread = new MessageListenerThread("kafka-listener", consumer, new MessageListener(null, null, logManager, 300_000L));
     }
 
     @Test
     void messages() throws IOException {
-        ConsumerRecord<String, byte[]> record = new ConsumerRecord<>("topic", 0, 1, "key", Strings.bytes("{}"));
-        record.headers().add(MessageHeaders.HEADER_CLIENT, Strings.bytes("client"));
-        record.headers().add(MessageHeaders.HEADER_REF_ID, Strings.bytes("refId"));
-        record.headers().add(MessageHeaders.HEADER_CORRELATION_ID, Strings.bytes("correlationId"));
+        var record = new ConsumerRecord<>("topic", 0, 1, "key", Strings.bytes("{}"));
+        record.headers().add(KafkaMessage.HEADER_CLIENT, Strings.bytes("client"));
+        record.headers().add(KafkaMessage.HEADER_REF_ID, Strings.bytes("refId"));
+        record.headers().add(KafkaMessage.HEADER_CORRELATION_ID, Strings.bytes("correlationId"));
         var actionLog = new ActionLog(null, null);
-        List<Message<TestMessage>> messages = thread.messages(List.of(record), actionLog, JSONMapper.reader(TestMessage.class));
+        List<Message<TestMessage>> messages = thread.messages(List.of(new KafkaMessage(record)), actionLog, JSONMapper.reader(TestMessage.class));
 
         assertThat(messages).hasSize(1);
         assertThat(messages.get(0).key).isEqualTo("key");
@@ -94,27 +85,27 @@ class MessageListenerThreadTest {
     }
 
     @Test
-    void handle() throws Exception {
+    void handleSingle() throws Exception {
         var key = "key";
         var message = new TestMessage();
         message.stringField = "value";
         var record = new ConsumerRecord<>("topic", 0, 0, key, Strings.bytes(JSON.toJSON(message)));
-        record.headers().add(MessageHeaders.HEADER_TRACE, Strings.bytes("true"));
-        record.headers().add(MessageHeaders.HEADER_CLIENT, Strings.bytes("client"));
-        thread.handle("topic", new MessageProcess<>(messageHandler, null, TestMessage.class), List.of(record));
+        record.headers().add(KafkaMessage.HEADER_TRACE, Strings.bytes("true"));
+        record.headers().add(KafkaMessage.HEADER_CLIENT, Strings.bytes("client"));
+        thread.handleSingle("topic", new MessageProcess<>(messageHandler, TestMessage.class), new KafkaMessage(record));
 
         verify(messageHandler).handle(eq(key), argThat(value -> "value".equals(value.stringField)));
     }
 
     @Test
-    void handleWithNullKey() throws Exception {
+    void handleSingleWithNullKey() throws Exception {
         var message = new TestMessage();
         message.stringField = "value";
         var record = new ConsumerRecord<>("topic", 0, 0, (String) null, Strings.bytes(JSON.toJSON(message)));
-        record.headers().add(MessageHeaders.HEADER_TRACE, Strings.bytes("true"));
-        record.headers().add(MessageHeaders.HEADER_CLIENT, Strings.bytes("client"));
-        record.headers().add(MessageHeaders.HEADER_TRACE, Strings.bytes("cascade"));
-        thread.handle("topic", new MessageProcess<>(messageHandler, null, TestMessage.class), List.of(record));
+        record.headers().add(KafkaMessage.HEADER_TRACE, Strings.bytes("true"));
+        record.headers().add(KafkaMessage.HEADER_CLIENT, Strings.bytes("client"));
+        record.headers().add(KafkaMessage.HEADER_TRACE, Strings.bytes("cascade"));
+        thread.handleSingle("topic", new MessageProcess<>(messageHandler, TestMessage.class), new KafkaMessage(record));
 
         verify(messageHandler).handle(isNull(), argThat(value -> "value".equals(value.stringField)));
     }
@@ -125,9 +116,9 @@ class MessageListenerThreadTest {
         var message = new TestMessage();
         message.stringField = "value";
         var record = new ConsumerRecord<>("topic", 0, 0, key, Strings.bytes(JSON.toJSON(message)));
-        record.headers().add(MessageHeaders.HEADER_CORRELATION_ID, Strings.bytes("correlationId"));
-        record.headers().add(MessageHeaders.HEADER_REF_ID, Strings.bytes("refId"));
-        thread.handleBulk("topic", new MessageProcess<>(null, bulkMessageHandler, TestMessage.class), List.of(record));
+        record.headers().add(KafkaMessage.HEADER_CORRELATION_ID, Strings.bytes("correlationId"));
+        record.headers().add(KafkaMessage.HEADER_REF_ID, Strings.bytes("refId"));
+        thread.handleBulk("topic", new MessageProcess<>(bulkMessageHandler, TestMessage.class), List.of(new KafkaMessage(record)));
 
         verify(bulkMessageHandler).handle(argThat(value -> value.size() == 1
                                                            && key.equals(value.get(0).key)